Лишние json и blob файлы не создаются.

+++ Bwl.Storage.UniversalORM.Blob/Savers/FileBlobSaver.vb
@@ -21,16 +21,19 @@ Public Class FileBlobSaver
 
 	Public Function Load(parentObjId As String) As ObjBlobInfo Implements IBlobSaver.Load
 		parentObjId = parentObjId.Replace(" ", "")
-		Dim dir = GetPath(parentObjId)
-		Dim json = File.ReadAllText(Path.Combine(dir, parentObjId + ".json"))
-		Dim objBlobInfo = JsonUtils.LoadFromJsonString(Of ObjBlobInfo)(json)
-		For Each blobInfo In objBlobInfo.BlobsInfo
-			blobInfo.Data = File.ReadAllBytes(Path.Combine(dir, blobInfo.BlobId))
-		Next
-		Return objBlobInfo
+		Dim dir = GetPath(parentObjId, True, False)
+		Dim objBlobInfo As ObjBlobInfo = Nothing
+		If Directory.Exists(dir) Then
+			Dim json = File.ReadAllText(Path.Combine(dir, parentObjId + ".json"))
+			objBlobInfo = JsonUtils.LoadFromJsonString(Of ObjBlobInfo)(json)
+			For Each blobInfo In objBlobInfo.BlobsInfo
+				blobInfo.Data = File.ReadAllBytes(Path.Combine(dir, blobInfo.BlobId))
+			Next
+		End If
+		Return ObjBlobInfo
 	End Function
 
-	Private Function GetPath(id As String, Optional fullPath As Boolean = True, Optional needCreate As Boolean = True)
+	Private Function GetPath(id As String, fullPath As Boolean, needCreate As Boolean)
 		Dim subDir = id
 		If subDir.Length > 10 Then
 			subDir = id.Substring(1, 8)
@@ -51,36 +54,40 @@ Public Class FileBlobSaver
 	End Function
 
 	Public Function GetBlobFilePath(id As String, blobName As String) As String
+		Dim res = String.Empty
 		id = id.Replace(" ", "")
-		Dim dir = GetPath(id)
-		Dim subDir = GetPath(id, False)
-		Dim json = File.ReadAllText(Path.Combine(dir, id + ".json"))
-		Dim objBlobInfo = JsonUtils.LoadFromJsonString(Of ObjBlobInfo)(json)
-		Dim bi = objBlobInfo.BlobsInfo.FirstOrDefault(Function(b) b.FieldName = blobName)
-		If (bi IsNot Nothing) Then
-			Return Path.Combine(subDir, bi.BlobId)
-		Else
-			Throw New Exception("FileblobSaver.GetBlobFilePath _ не найдено поле " + blobName)
+		Dim dir = GetPath(id, True, False)
+		Dim subDir = GetPath(id, False, False)
+		If Directory.Exists(dir) Then
+			Dim json = File.ReadAllText(Path.Combine(dir, id + ".json"))
+			Dim objBlobInfo = JsonUtils.LoadFromJsonString(Of ObjBlobInfo)(json)
+			Dim bi = objBlobInfo.BlobsInfo.FirstOrDefault(Function(b) b.FieldName = blobName)
+			If (bi IsNot Nothing) Then
+				res = Path.Combine(subDir, bi.BlobId)
+			Else
+				Throw New Exception("FileblobSaver.GetBlobFilePath _ не найдено поле " + blobName)
+			End If
 		End If
-		Return String.Empty
+		Return res
 	End Function
 
-
 	Public Sub Save(objBlobInfo As ObjBlobInfo) Implements IBlobSaver.Save
-		Dim dir = GetPath(objBlobInfo.ParentObjId)
-		Dim json = JsonUtils.ToJson(objBlobInfo)
-		File.WriteAllText(Path.Combine(dir, objBlobInfo.ParentObjId + ".json"), json)
+		If objBlobInfo IsNot Nothing AndAlso objBlobInfo.BlobsInfo IsNot Nothing AndAlso objBlobInfo.BlobsInfo.Any Then
+			Dim dir = GetPath(objBlobInfo.ParentObjId, True, True)
+			Dim json = JsonUtils.ToJson(objBlobInfo)
+			File.WriteAllText(Path.Combine(dir, objBlobInfo.ParentObjId + ".json"), json)
 
-		For Each blobInfo In objBlobInfo.BlobsInfo
-			Dim fname = Path.Combine(dir, blobInfo.BlobId)
-			If (File.Exists(fname)) Then
-				File.Delete(fname)
-			End If
-			Dim fileStream = New FileStream(fname, FileMode.CreateNew)
-			fileStream.Write(blobInfo.Data, 0, blobInfo.Data.Length)
-			fileStream.Close()
-			fileStream.Dispose()
-		Next
+			For Each blobInfo In objBlobInfo.BlobsInfo
+				Dim fname = Path.Combine(dir, blobInfo.BlobId)
+				If (File.Exists(fname)) Then
+					File.Delete(fname)
+				End If
+				Dim fileStream = New FileStream(fname, FileMode.CreateNew)
+				fileStream.Write(blobInfo.Data, 0, blobInfo.Data.Length)
+				fileStream.Close()
+				fileStream.Dispose()
+			Next
+		End If
 	End Sub
 
 	Sub Remove(parentObjId As String) Implements IBlobSaver.Remove
